chore: clean out all dynamic references + types
