+++ beet/src/beet.ts
@@ -20,6 +20,7 @@ export * from './beets/collections'
 export * from './beets/string'
 export * from './beets/composites'
 export * from './beets/numbers'
+export * from './beet.fixable'
 export * from './read-write'
 export * from './struct'
 export * from './struct.fixable'
